refactor(schedule-engine): Move core classes into 'classes' directoryRelocate Course, Schedule, Section, and SectionTime classes to adedicated 'classes' subdirectory. This improves module organization.
