+++ lib/schedule_engine/classes/course.ts
