+++ lib/schedule_engine/classes/section_time.ts
